server/scheduleScan: update twitch bot to v3 - fix id inputs
