+++ ui/.env.template
@@ -1,24 +1,35 @@
 # Example environment variables for development
 # Copy this to .env and fill in your actual values
 
-# Twitch Configuration
+# -----------------------------
+# Twitch Configuration (env names used by ui/src/config/settings.py)
+# -----------------------------
+# App credentials
 TWITCH_CLIENT_ID=
 TWITCH_CLIENT_SECRET=
-TWITCH_ACCESS_TOKEN=oauth:
-TWITCH_REFRESH_TOKEN=
+
+# Bot tokens (the bot account that connects and sends chat messages)
+# Include the 'oauth:' prefix on access tokens when required
+TWITCH_BOT_ID=896950964
+
+# General channel/owner settings
 TWITCH_CHANNEL=the8bitsaxlounge
+TWITCH_OWNER_ID=1424580736
+
+# Command prefix
 TWITCH_PREFIX=!
 
-# Bot Configuration
+# -----------------------------
+# Bot & logging
+# -----------------------------
 BOT_NAME=EightBitSaxBot
-BOT_OWNER_ID=896950964  # Twitch User ID for the bot owner (can be found via Twitch API)
-BOT_ID=896950964  # Twitch User ID for EightBitSaxBot (can be found via Twitch API)
 LOG_LEVEL=INFO
 
+# -----------------------------
 # MIDI Configuration
+# -----------------------------
 # Use host.containers.internal when running in Podman to access host services
 MIDI_DEVICE_URL=
-MIDI_DATA_URL=
 MIDI_API_TIMEOUT=30
 MIDI_CLIENT_ID=
 MIDI_CLIENT_SECRET=
