+++ ui/k8s/secret.yaml
@@ -11,9 +11,5 @@ type: Opaque
 stringData:
   client-id: "placeholder"
   client-secret: "placeholder"
-  bot-access-token: "placeholder"
-  bot-refresh-token: "placeholder"
-  channel-access-token: "placeholder"
-  channel-refresh-token: "placeholder"
   midi-client-id: "placeholder"
   midi-client-secret: "placeholder"
