+++ ui/src/config/settings.py
@@ -13,12 +13,8 @@ class Settings(BaseSettings):
     )
     
     # Twitch Configuration
-    twitch_bot_id: str  # Twitch User ID for EightBitSaxBot (can be found via Twitch API)
-    twitch_owner_id: str  # Twitch User ID for the bot owner
-    twitch_bot_access_token: str  # Access token with 'oauth:' prefix
-    twitch_bot_refresh_token: str  # Refresh token for obtaining new access tokens
-    twitch_channel_access_token: str  # Access token for the channel
-    twitch_channel_refresh_token: str  # Refresh token for the channel
+    twitch_bot_id: int  # Twitch User ID for EightBitSaxBot (can be found via Twitch API)
+    twitch_owner_id: int  # Twitch User ID for the bot owner
     twitch_client_id: str  # Your app's Client ID from dev.twitch.tv
     twitch_client_secret: str  # Your app's Client Secret from dev.twitch.tv
     twitch_channel: str  # Channel name to connect to
