+++ ui/src/main.py
@@ -1,46 +1,40 @@
-"""An example of connecting to a conduit and subscribing to EventSub when a User Authorizes the application.
-
-This bot can be restarted as many times without needing to subscribe or worry about tokens:
-- Tokens are stored in '.tio.tokens.json' by default
-- Subscriptions last 72 hours after the bot is disconnected and refresh when the bot starts.
-
-Consider reading through the documentation for AutoBot for more in depth explanations.
-"""
-
 import asyncio
 import logging
+import sys
 
-import asqlite
-import twitchio
-
+from bots.twitch.bot import Bot as StreamingBot
+from config.logging_config import configure_logging
+from config.settings import settings
 from services.health_server import HealthServer
-from bots.twitch.example_bot import Bot
-import bots.twitch.example_bot as eb
 
-LOGGER: logging.Logger = logging.getLogger("Bot")
+configure_logging(settings.log_level)
+logger = logging.getLogger(__name__)
 
-# Our main entry point for our Bot
-# Best to setup_logging here, before anything starts
-def main() -> None:
-    twitchio.utils.setup_logging(level=logging.INFO)
-
-    async def runner() -> None:
-        async with asqlite.create_pool("tokens.db") as tdb:
-            tokens, subs = await eb.setup_database(tdb)
-
-            async with Bot(token_database=tdb, subs=subs) as bot:
-                health_server = HealthServer(port=8080, bot_instance=bot)
-                await health_server.start()
-                for pair in tokens:
-                    await bot.add_token(*pair)
-
-                await bot.start(load_tokens=False)
+async def main() -> None:
+    """
+    Main function to run the bot and health server.
+    - Starts the health server first, then the implementation of StreamingBot
+    """
 
     try:
-        asyncio.run(runner())
-    except KeyboardInterrupt:
-        LOGGER.warning("Shutting down due to KeyboardInterrupt")
+        bot = StreamingBot()
+        health_server = HealthServer(port=8080, bot_instance=bot)
 
+        await health_server.start()
+        await bot.start()
+    except KeyboardInterrupt:
+        logger.warning("Shutting down due to KeyboardInterrupt")
+    except Exception as e:
+        logger.error(f'Fatal error: {e}')
+        await shutdown_all(bot, health_server)
+        sys.exit(1)
+    finally:
+        await shutdown_all(bot, health_server)
+
+async def shutdown_all(bot, health_server):
+    """Gracefully shutdown bot and health server."""
+    await bot.shutdown()
+    await health_server.stop()
 
 if __name__ == "__main__":
-    main()
+    asyncio.run(main())
