+++ ui/tests/conftest.py
@@ -14,10 +14,6 @@ import os
 import sys
 
 # Set environment variables before importing any app modules
-os.environ.setdefault('TWITCH_BOT_ACCESS_TOKEN', 'oauth:test_token')
-os.environ.setdefault('TWITCH_BOT_REFRESH_TOKEN', 'test_refresh_token')
-os.environ.setdefault('TWITCH_CHANNEL_ACCESS_TOKEN', 'oauth:channel_token')
-os.environ.setdefault('TWITCH_CHANNEL_REFRESH_TOKEN', 'channel_refresh_token')
 os.environ.setdefault('TWITCH_CLIENT_ID', 'test_client_id')
 os.environ.setdefault('TWITCH_CHANNEL', 'test_channel')
 os.environ.setdefault('MIDI_CLIENT_ID', 'test_midi_client')
@@ -31,10 +27,6 @@ sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', 'src'))
 def mock_settings():
     """Mock settings for testing."""
     settings = Mock()
-    settings.twitch_bot_access_token = "oauth:test_token"
-    settings.twitch_bot_refresh_token = "test_refresh_token"
-    settings.twitch_channel_access_token = "oauth:channel_token"
-    settings.twitch_channel_refresh_token = "channel_refresh_token"
     settings.twitch_client_id = "test_client_id"
     settings.twitch_channel = "test_channel"
     settings.twitch_prefix = "!"
