+++ ui/version.txt
@@ -1,1 +1,1 @@
-4.0.26
+4.0.27
